Move team themed entities to separate package
